requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
json2html==1.3.0
jsonpickle==4.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
openai==1.58.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
pyghidra==0.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
r2pipe==1.9.4
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
